Reverted because corruption of project organization (use latest Android Studio). Removed usage of deprecated methods.
I don't really know where that 700 came from, but it seems to work. I
think the core of the issue is the fact we have multiple children of a
FrameLayout (not good). We should change our parent layout to a Vertical
Linear Layout. I'll fix it this afternoon or just draw up what I'm
talking about. FrameLayouts generally should not have more than one
child.